+++ packages/applint/src/LinterImpl.ts
@@ -1,4 +1,0 @@
-export default interface LinterImpl {
-  scan: () => Promise<any>;
-  fix: () => Promise<any>;
-}