+++ packages/applint/src/index.ts
@@ -1,2 +1,4 @@
-export * from './eslint';
-export * from './projectLint';
+export * from './ESLint';
+export * from './Stylelint';
+export * from '@applint/projectlint';
+export type { RuleKey } from '@applint/spec';