+++ packages/applint/src/projectLint/index.ts
@@ -1,1 +1,0 @@
-export * from '@applint/projectlint';