+++ packages/spec/src/eslint/common-ts.ts
@@ -1,7 +1,10 @@
 import type { Linter } from 'eslint';
 
 const commonTypeScriptESLintConfig: Linter.Config = {
-  extends: ['@applint/eslint-config/typescript'],
+  extends: [
+    // if not use require.resolve(), the @applint/eslint-config can't be find in @appworks/doctor(call with Node API)
+    require.resolve('@applint/eslint-config/typescript'),
+  ],
 };
 
 export default commonTypeScriptESLintConfig;