+++ packages/applint/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @applint/applint
 
+## 1.0.0
+
+- feat: support scan and fix CSS files
+
 ## 0.1.1
 
 - feat: bump `@applint/spec` version to 1.0.0