+++ packages/spec/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @applint/spec
 
+## 1.1.0
+
+- feat: support lint vue files
+- fix: @typescript-eslint/parser dependency is lost
+- fix: can't find `@applint/eslint-config` in `@appworks/doctor`
+- chore: bump `stylelint-scss` version to `^4.0.0` to support `stylelint@^14`
+
 ## 1.0.1
 
 - feat: add `@applint/prettier-config`